Adding new JobGiver for PrisonLabor mod (Fixes #5)
PrisonLabor has a different method to have the prisoners do jobs,
therefor the existing JobGiver can't be used 1:1 for prisoners.
Adding a new one tailored to their needs. This should now also complete
the fix for issue 5 that was reopened, as unoccupied prisoners just went
drinking water again and again.